refactor: modularize plan validator
