+++ agent_s3/tools/plan_validator/config.py
@@ -1,0 +1,168 @@
+"""Configuration constants for plan validation."""
+
+from __future__ import annotations
+
+import keyword
+
+# Maximum token counts for different complexity levels
+TOKEN_BUDGET_LIMITS = {
+    0: 2000,  # trivial
+    1: 5000,  # simple
+    2: 10000,  # moderate
+    3: 20000,  # complex
+}
+
+# Language-specific keywords
+PYTHON_KEYWORDS = set(keyword.kwlist)
+JS_KEYWORDS = {
+    "break",
+    "case",
+    "catch",
+    "class",
+    "const",
+    "continue",
+    "debugger",
+    "default",
+    "delete",
+    "do",
+    "else",
+    "export",
+    "extends",
+    "false",
+    "finally",
+    "for",
+    "function",
+    "if",
+    "import",
+    "in",
+    "instanceof",
+    "new",
+    "null",
+    "return",
+    "super",
+    "switch",
+    "this",
+    "throw",
+    "true",
+    "try",
+    "typeof",
+    "var",
+    "void",
+    "while",
+    "with",
+    "yield",
+    # ES6+ keywords
+    "let",
+    "static",
+    "await",
+    "async",
+}
+PHP_KEYWORDS = {
+    "__halt_compiler",
+    "abstract",
+    "and",
+    "array",
+    "as",
+    "break",
+    "callable",
+    "case",
+    "catch",
+    "class",
+    "clone",
+    "const",
+    "continue",
+    "declare",
+    "default",
+    "die",
+    "do",
+    "echo",
+    "else",
+    "elseif",
+    "empty",
+    "enddeclare",
+    "endfor",
+    "endforeach",
+    "endif",
+    "endswitch",
+    "endwhile",
+    "eval",
+    "exit",
+    "extends",
+    "final",
+    "finally",
+    "fn",
+    "for",
+    "foreach",
+    "function",
+    "global",
+    "goto",
+    "if",
+    "implements",
+    "include",
+    "include_once",
+    "instanceof",
+    "insteadof",
+    "interface",
+    "isset",
+    "list",
+    "namespace",
+    "new",
+    "or",
+    "print",
+    "private",
+    "protected",
+    "public",
+    "require",
+    "require_once",
+    "return",
+    "static",
+    "switch",
+    "throw",
+    "trait",
+    "try",
+    "unset",
+    "use",
+    "var",
+    "while",
+    "xor",
+    "yield",
+}
+
+# System-reserved environment variable names
+RESERVED_ENV_VARS = {
+    "PATH",
+    "HOME",
+    "USER",
+    "SHELL",
+    "LANG",
+    "TERM",
+    "EDITOR",
+    "PYTHONPATH",
+    "JAVA_HOME",
+    "LD_LIBRARY_PATH",
+    "TEMP",
+    "TMP",
+    "NODE_ENV",
+    "PORT",
+    "HOST",
+    "DATABASE_URL",
+    "PWD",
+    "PS1",
+    "DISPLAY",
+    "HOSTNAME",
+    "LOGNAME",
+    "_",
+    "OLDPWD",
+    "MAIL",
+    "SSH_CONNECTION",
+    "SSH_TTY",
+    "TERM_PROGRAM",
+    "SHELL_SESSION_ID",
+    "HTTP_HOST",
+    "DOCUMENT_ROOT",
+    "SERVER_PROTOCOL",
+    "REQUEST_METHOD",
+    "SCRIPT_FILENAME",
+    "QUERY_STRING",
+    "PHP_SELF",
+}
