+++ agent_s3/tools/plan_validator/structural_checks.py
@@ -1,483 +1,9 @@
-"""Static Plan Checker to validate Pre-Planner JSON output.
+from __future__ import annotations
 
-This module provides fast, deterministic validation of Pre-Planner outputs
-before they reach the Planner phase. It verifies structural correctness,
-identifier uniqueness, path validity, token budget compliance, and other
-constraints without requiring LLM calls.
+from typing import Any, Dict, List
 
-Implementation uses comprehensive AST parsing for multi-language code analysis.
-"""
-
-import json
-import re
-import os
-import logging
-import keyword
-import ast
-import xml.etree.ElementTree as ET
-import datetime
-from typing import Dict, Any, List, Tuple
-from collections import defaultdict
-from dataclasses import dataclass
-from pathlib import Path
-
-# Import AST tools for multi-language support
-from agent_s3.ast_tools.parser import parse_js
-from .parsing.parser_registry import ParserRegistry
-
-logger = logging.getLogger(__name__)
-
-# Maximum token counts for different complexity levels
-TOKEN_BUDGET_LIMITS = {
-    0: 2000,  # trivial
-    1: 5000,  # simple
-    2: 10000, # moderate
-    3: 20000  # complex
-}
-
-# Language-specific keywords
-PYTHON_KEYWORDS = set(keyword.kwlist)
-JS_KEYWORDS = {
-    "break", "case", "catch", "class", "const", "continue", "debugger",
-    "default", "delete", "do", "else", "export", "extends", "false",
-    "finally", "for", "function", "if", "import", "in", "instanceof",
-    "new", "null", "return", "super", "switch", "this", "throw", "true",
-    "try", "typeof", "var", "void", "while", "with", "yield",
-    # ES6+ keywords
-    "let", "static", "await", "async"
-}
-PHP_KEYWORDS = {
-    "__halt_compiler", "abstract", "and", "array", "as", "break",
-    "callable", "case", "catch", "class", "clone", "const", "continue",
-    "declare", "default", "die", "do", "echo", "else", "elseif",
-    "empty", "enddeclare", "endfor", "endforeach", "endif", "endswitch",
-    "endwhile", "eval", "exit", "extends", "final", "finally", "fn",
-    "for", "foreach", "function", "global", "goto", "if", "implements",
-    "include", "include_once", "instanceof", "insteadof", "interface",
-    "isset", "list", "namespace", "new", "or", "print", "private",
-    "protected", "public", "require", "require_once", "return", "static",
-    "switch", "throw", "trait", "try", "unset", "use", "var", "while", "xor", "yield"
-}
-
-# System-reserved environment variable names
-RESERVED_ENV_VARS = {
-    "PATH", "HOME", "USER", "SHELL", "LANG", "TERM", "EDITOR",
-    "PYTHONPATH", "JAVA_HOME", "LD_LIBRARY_PATH", "TEMP", "TMP",
-    "NODE_ENV", "PORT", "HOST", "DATABASE_URL", "PWD", "PS1",
-    "DISPLAY", "HOSTNAME", "LOGNAME", "_", "OLDPWD", "MAIL",
-    "SSH_CONNECTION", "SSH_TTY", "TERM_PROGRAM", "SHELL_SESSION_ID",
-    "HTTP_HOST", "DOCUMENT_ROOT", "SERVER_PROTOCOL", "REQUEST_METHOD",
-    "SCRIPT_FILENAME", "QUERY_STRING", "PHP_SELF"
-}
-
-
-class PlanValidationError(Exception):
-    """Exception raised when plan validation fails."""
-    pass
-
-
-@dataclass
-class CodeElement:
-    """Represents a code element with position information."""
-    name: str
-    element_type: str
-    start_line: int = 0
-    end_line: int = 0
-    params: List[str] = None
-
-    def __post_init__(self):
-        if self.params is None:
-            self.params = []
-
-    def __repr__(self) -> str:
-        """String representation of the element."""
-        return f"{self.element_type}:{self.name}:{self.start_line}-{self.end_line}"
-
-
-@dataclass
-class SemanticRelation:
-    """Represents a semantic relationship between code elements."""
-    source: str
-    target: str
-    relation_type: str
-
-    def __repr__(self) -> str:
-        """String representation of the relationship."""
-        return f"{self.source} {self.relation_type} {self.target}"
-
-
-class CodeAnalyzer:
-    """AST-based code analyzer for extracting code elements across languages."""
-    def __init__(self, parser_registry: ParserRegistry = None, file_tool=None):
-        self.parser_registry = parser_registry or ParserRegistry()
-        self.file_tool = file_tool
-
-    def analyze_code(
-        self,
-        code_str: str,
-        lang: str | None = None,
-        file_path: str | None = None,
-        tech_stack: dict | None = None,
-    ) -> Dict[str, Any]:
-        """Analyze code and extract semantic elements using the pluggable parsers."""
-
-        # Auto-detect language if not provided
-        if not lang and file_path:
-            parser = self.parser_registry.get_parser(file_path=file_path)
-            if parser:
-                lang = parser.__class__.__name__.replace("Parser", "").lower()
-
-        if not lang:
-            lang = self._detect_language(code_str)
-
-        parser = self.parser_registry.get_parser(file_path=file_path, language_name=lang)
-        if parser:
-            return parser.analyze(code_str, file_path or "", tech_stack)
-
-        logger.error("No parser found for language '%s'. Skipping analysis.", lang)
-        return {
-            "language": lang,
-            "elements": [],
-            "relations": [],
-            "imports": set(),
-            "functions": set(),
-            "classes": set(),
-            "variables": set(),
-            "route_paths": set(),
-            "env_vars": set(),
-        }
-
-    def _analyze_file(self, file_path: str, language: str = None):
-        """
-        Analyze a file using the new parser system only.
-        """
-        if not self.file_tool:
-            logger.error("FileTool is not available. Cannot analyze file.")
-            return None
-        if not self.parser_registry:
-            logger.error("ParserRegistry not available. Cannot analyze file.")
-            return None
-        try:
-            content = self.file_tool.read_file(file_path)
-            if content is None:
-                logger.warning(
-                    "Could not read content of file: %s",
-                    file_path,
-                )
-                return None
-        except Exception as e:
-            logger.error(f"Error reading file {file_path}: {e}", exc_info=True)
-            return None
-        # Language detection
-        if not language:
-            if hasattr(self.file_tool, 'get_language_from_extension'):
-                language = self.file_tool.get_language_from_extension(file_path)
-            if not language:
-                language = Path(file_path).suffix[1:].lower() if Path(file_path).suffix else 'unknown'
-        parser = self.parser_registry.get_parser(language_name=language, file_path=file_path)
-        if parser:
-            try:
-                logger.info(
-                    "Analyzing %s with %s for language '%s'.",
-                    file_path,
-                    type(parser).__name__,
-                    language,
-                )
-                structure = parser.parse_code(content, file_path)
-                return structure
-            except Exception as e:
-                logger.error(f"Error analyzing {file_path} with {type(parser).__name__}: {e}", exc_info=True)
-                return None
-        else:
-            logger.error(
-                "No parser found for language '%s' for file %s.",
-                language,
-                file_path,
-            )
-            return None
-
-# Main validation functions
-def validate_pre_plan(
-    data: Dict[str, Any],
-    repo_root: str | None = None,
-    context_registry=None,
-) -> Tuple[bool, Dict[str, Any]]:
-    """
-    Validate Pre-Planner JSON output with fast, deterministic checks.
-
-    This function performs comprehensive validation of the pre-planning JSON output,
-    including schema validation, code syntax checking, identifier hygiene, path validity,
-    token budget compliance, duplicate symbol detection, and content scanning for
-    dangerous operations.
-
-    Args:
-        data: The Pre-Planner JSON data structure to validate
-        repo_root: Optional repository root path for file glob validation
-        context_registry: Optional context registry for accessing project context
-
-    Returns:
-        Tuple of (is_valid, validation_results dict with critical errors and warnings)
-    """
-    if repo_root is None:
-        repo_root = os.getcwd()
-
-    # Initialize structured validation results
-    validation_results = {
-        "critical": [],  # Critical errors that should block workflow
-        "warnings": [],  # Warnings that can be addressed but don't block
-        "suggestions": [],  # Optimization suggestions
-        "sections": {
-            "architecture": False,
-            "implementation": False,
-            "tests": False
-        }
-    }
-
-    # Scan for dangerous content patterns
-    dangerous_patterns = [
-        "rm -rf", "deltree", "format", "DROP TABLE", "DROP DATABASE",
-        "DELETE FROM", "TRUNCATE TABLE", "sudo", "chmod 777",
-        "eval(", "exec(", "system(", "shell_exec", "os.system"
-    ]
-
-    # Check for dangerous content in descriptions, signatures, etc.
-    for group_idx, group in enumerate(data.get("feature_groups", [])):
-        if not isinstance(group, dict):
-            continue
-
-        for feature_idx, feature in enumerate(group.get("features", [])):
-            if not isinstance(feature, dict):
-                continue
-
-            # Check feature description for dangerous content
-            description = feature.get("description", "")
-            for pattern in dangerous_patterns:
-                if pattern.lower() in description.lower():
-                    validation_results["critical"].append({
-                        "message": f"Feature '{feature.get('name', f'at index {feature_idx}')}' contains potentially dangerous operation '{pattern}' in description",
-                        "category": "security",
-                        "suggestion": f"Remove or replace the dangerous operation '{pattern}'"
-                    })
-
-            # Check system_design elements
-            if "system_design" in feature and isinstance(feature["system_design"], dict):
-                # Mark architecture section as present
-                validation_results["sections"]["architecture"] = True
-
-                # Check code elements
-                for el_idx, element in enumerate(feature["system_design"].get("code_elements", [])):
-                    if not isinstance(element, dict):
-                        continue
-
-                    # Check signature and description
-                    for field in ["signature", "description"]:
-                        content = element.get(field, "")
-                        for pattern in dangerous_patterns:
-                            if pattern.lower() in content.lower():
-                                validation_results["critical"].append({
-                                    "message": f"Code element '{element.get('name', f'at index {el_idx}')}' contains potentially dangerous operation '{pattern}' in {field}",
-                                    "category": "security",
-                                    "suggestion": f"Remove or replace the dangerous operation '{pattern}'"
-                                })
-
-            # Check test_requirements
-            if "test_requirements" in feature and isinstance(feature["test_requirements"], dict):
-                # Mark tests section as present
-                validation_results["sections"]["tests"] = True
-
-                # Ensure at least one test case exists across all test types
-                test_req = feature["test_requirements"]
-                total_tests = 0
-                for tests in test_req.values():
-                    if isinstance(tests, list):
-                        total_tests += len(tests)
-
-                if total_tests == 0:
-                    validation_results["critical"].append({
-                        "message": (
-                            f"Feature '{feature.get('name', f'at index {feature_idx}')}' "
-                            "must include at least one test case in 'test_requirements'"
-                        ),
-                        "category": "tests",
-                        "suggestion": "Add unit, integration, property-based, or acceptance tests"
-                    })
-
-            # Check implementation steps if present
-            if "implementation_steps" in feature and isinstance(feature["implementation_steps"], list):
-                valid_steps_found = False
-                for step_idx, step in enumerate(feature["implementation_steps"]):
-                    if not isinstance(step, dict):
-                        continue
-                    valid_steps_found = True
-
-                    # Check step description and code for dangerous content
-                    for field in ["description", "code"]:
-                        content = step.get(field, "")
-                        for pattern in dangerous_patterns:
-                            if pattern.lower() in content.lower():
-                                validation_results["critical"].append({
-                                    "message": f"Implementation step {step_idx} contains potentially dangerous operation '{pattern}' in {field}",
-                                    "category": "security",
-                                    "suggestion": f"Remove or replace the dangerous operation '{pattern}'"
-                                })
-
-                if valid_steps_found:
-                    validation_results["sections"]["implementation"] = True
-
-    # Run all validation checks with severity classification
-    # Schema validation - critical if basic structure is invalid
-    for error in validate_schema(data):
-        if "missing required field" in error.lower() or "must be a" in error.lower():
-            validation_results["critical"].append({
-                "message": error,
-                "category": "schema",
-                "suggestion": None
-            })
-        else:
-            validation_results["warnings"].append({
-                "message": error,
-                "category": "schema",
-                "suggestion": None
-            })
-
-    # Code syntax validation - critical if code stubs have syntax errors
-    for error in validate_code_syntax(data):
-        validation_results["critical"].append({
-            "message": error["message"],
-            "category": "syntax",
-            "suggestion": error.get("suggestion")
-        })
-
-    # Identifier hygiene validation - critical if reserved keywords used
-    for error in validate_identifier_hygiene(data):
-        if "reserved" in error.lower() or "duplicate" in error.lower():
-            validation_results["critical"].append({
-                "message": error,
-                "category": "identifiers",
-                "suggestion": None
-            })
-        else:
-            validation_results["warnings"].append({
-                "message": error,
-                "category": "identifiers",
-                "suggestion": None
-            })
-
-    # Path validity - warnings only
-    for error in validate_path_validity(data, repo_root):
-        validation_results["warnings"].append({
-            "message": error,
-            "category": "paths",
-            "suggestion": None
-        })
-
-    # Token budget - critical if significantly over budget
-    for error in validate_token_budget(data):
-        if "exceeds global budget" in error.lower():
-            validation_results["critical"].append({
-                "message": error,
-                "category": "tokens",
-                "suggestion": "Consider breaking down the task into smaller subtasks"
-            })
-        else:
-            validation_results["warnings"].append({
-                "message": error,
-                "category": "tokens",
-                "suggestion": None
-            })
-
-    # Duplicate symbols - critical error
-    for error in validate_duplicate_symbols(data):
-        validation_results["critical"].append({
-            "message": error,
-            "category": "duplicates",
-            "suggestion": None
-        })
-
-    # Reserved prefixes - warnings
-    for error in validate_reserved_prefixes(data):
-        validation_results["warnings"].append({
-            "message": error,
-            "category": "env_vars",
-            "suggestion": None
-        })
-
-    # Stub test coherence - warnings
-    for error in validate_stub_test_coherence(data):
-        validation_results["warnings"].append({
-            "message": error,
-            "category": "tests",
-            "suggestion": None
-        })
-
-    # Complexity sanity - warnings
-    for error in validate_complexity_sanity(data):
-        validation_results["warnings"].append({
-            "message": error,
-            "category": "complexity",
-            "suggestion": None
-        })
-
-    # Check for missing sections
-    missing_sections = []
-    for section, present in validation_results["sections"].items():
-        if not present:
-            missing_sections.append(section)
-            validation_results["critical"].append({
-                "message": f"Missing required section: {section.capitalize()}",
-                "category": "completeness",
-                "suggestion": f"Add the {section.capitalize()} section to ensure a complete plan"
-            })
-
-    # Calculate summary counts
-    validation_results["summary"] = {
-        "critical_count": len(validation_results["critical"]),
-        "warning_count": len(validation_results["warnings"]),
-        "suggestion_count": len(validation_results["suggestions"]),
-        "missing_sections": missing_sections
-    }
-
-    # Plan is invalid if there are any critical errors
-    is_valid = len(validation_results["critical"]) == 0
-
-    # Add structured errors to context registry if available
-    if context_registry and not is_valid:
-        try:
-            context_registry.add("validation_errors", validation_results)
-        except Exception:
-            # Don't fail validation if context registry update fails
-            pass
-
-    # Generate a more detailed error report
-    if not is_valid:
-        error_report = {
-            "timestamp": datetime.datetime.now().isoformat(),
-            "validation_status": "failed",
-            "critical_errors": validation_results["critical"],
-            "warnings": validation_results["warnings"],
-            "suggestions": validation_results["suggestions"],
-            "sections_status": {
-                section: "missing" if section in missing_sections else "present"
-                for section in validation_results["sections"]
-            }
-        }
-
-        # Save error report to file for debugging
-        try:
-            report_path = "validation_error_report.json"
-            with open(report_path, "w", encoding="utf-8") as f:
-                json.dump(error_report, f, indent=2)
-            logger.info(
-                "Validation error report saved to %s",
-                report_path,
-            )
-        except Exception as e:
-            logger.error(
-                "Failed to save validation error report: %s",
-                e,
-            )
-
-    return is_valid, validation_results
+from .ast_utils import CodeAnalyzer
+from .config import JS_KEYWORDS, PHP_KEYWORDS, PYTHON_KEYWORDS, RESERVED_ENV_VARS
 
 
 def validate_schema(data: Dict[str, Any]) -> List[str]:
@@ -674,6 +200,8 @@ def validate_schema(data: Dict[str, Any]) -> List[str]:
     return errors
 
 
+
+def validate_identifier_hygiene(data: Dict[str, Any]) -> List[str]:
 def validate_identifier_hygiene(data: Dict[str, Any]) -> List[str]:
     """
     Validate identifier hygiene (naming conventions, keywords) in system_design.code_elements.
@@ -765,6 +293,7 @@ def validate_identifier_hygiene(data: Dict[str, Any]) -> List[str]:
     return errors
 
 
+
 def validate_path_validity(data: Dict[str, Any], repo_root: str) -> List[str]:
     """Validate that specified file paths are syntactically valid, relative, and do not traverse upwards."""
     errors = []
@@ -853,43 +382,6 @@ def validate_path_validity(data: Dict[str, Any], repo_root: str) -> List[str]:
     return errors
 
 
-def validate_token_budget(data: Dict[str, Any]) -> List[str]:
-    """Validate that token budget estimates are within limits."""
-    errors = []
-
-    # Track total token usage across all feature groups
-    total_tokens = 0
-
-    for i, group in enumerate(data.get("feature_groups", [])):
-        if not isinstance(group, dict):
-            continue
-
-        group_tokens = 0
-
-        for j, feature in enumerate(group.get("features", [])):
-            if not isinstance(feature, dict):
-                continue
-
-            # Extract token estimate and complexity
-            est_tokens = feature.get("est_tokens", 0)
-            complexity = feature.get("complexity_enum", 0)
-
-            # Validate tokens against complexity
-            if complexity in TOKEN_BUDGET_LIMITS and est_tokens > TOKEN_BUDGET_LIMITS[complexity]:
-                errors.append(f"Feature {feature.get('name', f'{j}')} in group {group.get('group_name', f'{i}')} " +
-                                                                                         f"exceeds token budget: {est_tokens} tokens for complexity level {complexity}")
-
-            group_tokens += est_tokens
-
-        total_tokens += group_tokens
-
-    # Check against global token budget (if defined)
-    global_budget = data.get("token_budget", 100000)
-    if total_tokens > global_budget:
-        errors.append(f"Total token estimate ({total_tokens}) exceeds global budget ({global_budget})")
-
-    return errors
-
 
 def validate_duplicate_symbols(data: Dict[str, Any]) -> List[str]:
     """Check for duplicate symbols within each feature's system_design.code_elements and optionally across features."""
@@ -968,6 +460,7 @@ def validate_duplicate_symbols(data: Dict[str, Any]) -> List[str]:
                         f"{feature_location_log_prefix}: Duplicate symbol '{s_name}' in file '{s_file}' defined {len(locations)} times at: {locations_str} within this feature's system_design."
                     )
     return errors
+
 
 
 def validate_reserved_prefixes(data: Dict[str, Any]) -> List[str]:
@@ -1056,6 +549,7 @@ def validate_reserved_prefixes(data: Dict[str, Any]) -> List[str]:
         #     errors.append(f"Environment variable '{env_var}' should use single underscore separator")
 
     return errors
+
 
 
 def validate_stub_test_coherence(data: Dict[str, Any]) -> List[str]:
@@ -1268,267 +762,4 @@ def validate_stub_test_coherence(data: Dict[str, Any]) -> List[str]:
     return errors
 
 
-def validate_code_syntax(data: Dict[str, Any]) -> List[Dict[str, Any]]:
-    """
-    Validate syntax of code signatures in system_design.code_elements.
 
-    Args:
-        data: Parsed JSON data from pre-planner
-
-    Returns:
-        List of error objects, each with 'feature_name', 'group_idx', 'feature_idx', 'element_idx', 'element_name', 'message'
-    """
-    errors = []
-    # analyzer = CodeAnalyzer() # Not strictly needed if only parsing signatures with ast/tree-sitter directly
-
-    for group_idx, group_data in enumerate(data.get("feature_groups", [])):
-        if not isinstance(group_data, dict):
-            continue
-        for feature_idx, feature_data in enumerate(group_data.get("features", [])):
-            if not isinstance(feature_data, dict):
-                continue
-
-            current_feature_name = feature_data.get("name", f"Feature {feature_idx}")
-
-            system_design = feature_data.get("system_design")
-            if not isinstance(system_design, dict):
-                # errors.append({...}) # Covered by schema validation
-                continue
-
-            code_elements = system_design.get("code_elements")
-            if not isinstance(code_elements, list):
-                # errors.append({...}) # Covered by schema validation
-                continue
-
-            for el_idx, element in enumerate(code_elements):
-                if not isinstance(element, dict):
-                    # errors.append({...}) # Covered by schema validation
-                    continue
-
-                signature_str = element.get("signature")
-                element_type = element.get("element_type")
-                element_name = element.get("name", f"unnamed_element_{el_idx}")
-                target_file = element.get("target_file", "")
-
-                if not isinstance(signature_str, str) or not signature_str.strip():
-                    errors.append({
-                        "feature_name": current_feature_name, "group_idx": group_idx, "feature_idx": feature_idx,
-                        "element_idx": el_idx, "element_name": element_name,
-                        "message": "Missing or empty 'signature' for code_element."
-                    })
-                    continue
-
-                # Determine language for parsing
-                lang = "python" # Default
-                if isinstance(target_file, str):
-                    if target_file.endswith((".js", ".jsx", ".mjs")):
-                        lang = "javascript"
-                    elif target_file.endswith((".ts", ".tsx")):
-                        lang = "typescript"
-                    elif target_file.endswith(".php"):
-                        lang = "php"
-                    elif target_file.endswith(".java"):
-                        lang = "java"
-
-                parse_content = signature_str
-                try:
-                    if lang == "python":
-                        # Attempt to make it a parsable snippet
-                        if element_type == "function" and signature_str.strip().endswith(":"):
-                            parse_content = f"{signature_str}\n  pass"
-                        elif element_type == "class" and signature_str.strip().endswith("):"):
-                             parse_content = f"{signature_str}\n  pass"
-                        elif element_type == "class" and not signature_str.strip().endswith("):") and ":" not in signature_str : # e.g. "class MyClass"
-                             parse_content = f"{signature_str}:\n  pass"
-
-                        ast.parse(parse_content)
-
-                    elif lang in ["javascript", "typescript"]:
-                        # Attempt to make it a parsable snippet for tree-sitter
-                        if element_type == "function":
-                            # Handle arrow functions like `(a: type) => type:` or `name = (a:type): type =>`
-                            if "=>" in signature_str:
-                                if not signature_str.strip().endswith(";") and not signature_str.strip().endswith("}"):
-                                     # Try to wrap in a const assignment if it looks like an arrow func expression
-                                     if not re.match(r"^\s*(const|let|var)\s+\w+
-                                         \s*=", signature_str):                                         parse_content = f"const tempFunc = {signature_str};"
-                                     elif not signature_str.strip().endswith(";"):
-                                         parse_content = f"{signature_str};"
-
-                            elif '(' in signature_str and ')' in signature_str and not signature_str.strip().endswith(";") and not signature_str.strip().endswith("}"):
-                                parse_content = f"{signature_str} {{}}" # For `function name(args): type`
-                        elif element_type == "class" and not signature_str.strip().endswith("}"):
-                             parse_content = f"{signature_str} {{}}"
-                        elif element_type == "interface" and not signature_str.strip().endswith("}"):
-                             parse_content = f"{signature_str} {{}}"
-
-                        # Use the appropriate parser from agent_s3.ast_tools
-                        if lang == "javascript":
-                            parse_js(bytes(parse_content, "utf8"))
-                        elif lang == "typescript":
-                            # from agent_s3.ast_tools.parser import parse_ts # Ensure this is available
-                            # For now, using parse_js as a fallback for basic TS syntax if parse_ts isn't set up
-                            # This is a simplification; a real TS parser would be better.
-                            try:
-                                from agent_s3.ast_tools.ts_languages import get_ts_parser # Try to import specific TS parser
-                                parser = get_ts_parser()
-                                parser.parse(bytes(parse_content, "utf8"))
-                            except ImportError:
-                                logger.warning("TypeScript parser not fully available, using JavaScript parser for basic syntax check of TS.")
-                                parse_js(bytes(parse_content, "utf8"))
-
-
-                    elif lang == "php":
-                        # PHP parsing is complex. A simple check:
-                        if element_type == "function" and "function " not in signature_str:
-                             errors.append({"message": f"PHP function signature for '{element_name}' should start with 'function '."}) # Example
-                        # Full PHP syntax check usually requires `php -l` or a dedicated PHP parser lib
-                        pass # Placeholder for more robust PHP check
-
-                    elif lang == "java":
-                        # Java parsing is also complex. Basic check:
-                        if element_type == "class" and "class " not in signature_str:
-                            errors.append({"message": f"Java class signature for '{element_name}' should start with 'class '."})
-                        elif element_type == "interface" and "interface " not in signature_str:
-                            errors.append({"message": f"Java interface signature for '{element_name}' should start with 'interface '."})
-                        # Full Java syntax check requires a Java compiler/parser
-                        pass # Placeholder
-
-                except SyntaxError as e: # Python's ast.parse specific error
-                    errors.append({
-                        "feature_name": current_feature_name, "group_idx": group_idx, "feature_idx": feature_idx,
-                        "element_idx": el_idx, "element_name": element_name,
-                        "message": f"Python syntax error in signature '{signature_str[:50]}...': {e.msg} (line {e.lineno}, offset {e.offset}). Parsed as: '{parse_content[:50]}...'"
-                    })
-                except Exception as e: # Generic error for other parsers or issues
-                    errors.append({
-                        "feature_name": current_feature_name, "group_idx": group_idx, "feature_idx": feature_idx,
-                        "element_idx": el_idx, "element_name": element_name,
-                        "message": f"Syntax error or parsing issue in {lang} signature '{signature_str[:50]}...': {str(e)}. Parsed as: '{parse_content[:50]}...'"
-                    })
-    return errors
-
-
-
-def validate_complexity_sanity(data: Dict[str, Any]) -> List[str]:
-    """Check that complexity levels correlate with token estimates."""
-    errors = []
-
-    for i, group in enumerate(data.get("feature_groups", [])):
-        if not isinstance(group, dict):
-            continue
-
-        for j, feature in enumerate(group.get("features", [])):
-            if not isinstance(feature, dict):
-                continue
-
-            # Extract token estimate and complexity
-            est_tokens = feature.get("est_tokens", 0)
-            complexity = feature.get("complexity_enum", 0)
-
-            # Check if complexity level exists
-            if complexity not in TOKEN_BUDGET_LIMITS:
-                errors.append(f"Feature {feature.get('name', f'{j}')} in group {group.get('group_name', f'{i}')} " +
-                                                                                         f"has invalid complexity level: {complexity}")
-                continue
-
-            # Sanity checks for complexity vs. token count
-            if complexity == 0 and est_tokens > TOKEN_BUDGET_LIMITS[0]:  # trivial
-                errors.append(f"Feature {feature.get('name', f'{j}')} in group {group.get('group_name', f'{i}')} " +
-                                                                                         f"marked as trivial but has {est_tokens} tokens (expected < {TOKEN_BUDGET_LIMITS[0]})")
-            elif complexity == 1 and est_tokens > TOKEN_BUDGET_LIMITS[1]:  # simple
-                errors.append(f"Feature {feature.get('name', f'{j}')} in group {group.get('group_name', f'{i}')} " +
-                                                                                         f"marked as simple but has {est_tokens} tokens (expected < {TOKEN_BUDGET_LIMITS[1]})")
-            elif complexity == 2 and est_tokens > TOKEN_BUDGET_LIMITS[2]:  # moderate
-                errors.append(f"Feature {feature.get('name', f'{j}')} in group {group.get('group_name', f'{i}')} " +
-                                                                                         f"marked as moderate but has {est_tokens} tokens (expected < {TOKEN_BUDGET_LIMITS[2]})")
-
-    return errors
-
-
-def write_junit_report(errors: List[str], output_path: str = 'plan_validation.xml') -> bool:
-    """
-    Write validation errors as JUnit XML for CI integration.
-
-    Args:
-        errors: List of validation error messages
-        output_path: Path to write the JUnit XML report
-
-    Returns:
-        True if report was written successfully
-    """
-    try:
-        # Create the root element
-        testsuites = ET.Element("testsuites")
-
-        # Create a testsuite element
-        testsuite = ET.SubElement(testsuites, "testsuite")
-        testsuite.set("name", "Plan Validation")
-        testsuite.set("tests", str(len(errors) if errors else 1))
-        testsuite.set("errors", str(len(errors)))
-        testsuite.set("failures", "0")
-        testsuite.set("timestamp", datetime.datetime.now().isoformat())
-
-        if errors:
-            # Create a testcase for each error
-            for error in errors:
-                testcase = ET.SubElement(testsuite, "testcase")
-                testcase.set("name", error[:40] + "..." if len(error) > 40 else error)
-                testcase.set("classname", "plan_validator")
-
-                # Add error element
-                error_elem = ET.SubElement(testcase, "error")
-                error_elem.set("message", error)
-                error_elem.text = error
-        else:
-            # Add a single successful test case
-            testcase = ET.SubElement(testsuite, "testcase")
-            testcase.set("name", "Plan validation passed")
-            testcase.set("classname", "plan_validator")
-
-        # Create the XML tree and write to file
-        tree = ET.ElementTree(testsuites)
-        tree.write(output_path, encoding="utf-8", xml_declaration=True)
-
-        return True
-    except Exception as e:
-        logger.error("Error writing JUnit report: %s", e)
-        return False
-
-
-def create_github_annotations(errors: List[str]) -> List[Dict[str, Any]]:
-    """
-    Format validation errors as GitHub annotations.
-
-    Args:
-        errors: List of validation error messages
-
-    Returns:
-        List of annotation dictionaries
-    """
-    annotations = []
-
-    for error in errors:
-        annotation = {
-            "message": error,
-            "annotation_level": "failure",
-            "title": "Plan Validation Error"
-        }
-
-        # Try to extract file path and line number if available
-        file_match = re.search(r'in\s+file\s+[\'"]?([^\'"\s]+)[\'"]?', error)
-        if file_match:
-            annotation["path"] = file_match.group(1)
-
-        # Try to extract specific path for duplicate route paths
-        route_match = re.search(r'Route\s+[\'"]([^\'"]+)[\'"]', error)
-        if route_match:
-            annotation["path"] = "api_routes.md"  # Create a virtual file for route issues
-
-        # Mark as warning for certain types of issues
-        if any(x in error.lower() for x in ["warning", "suggest", "recommend", "consider"]):
-            annotation["annotation_level"] = "warning"
-
-        annotations.append(annotation)
-
-    return annotations
